update sirun benchmark variants to run in parallel (#3009)
