+++ benchmark/sirun/appsec/common.js
@@ -1,6 +1,6 @@
 'use strict'
 
 module.exports = {
-  port: 3032,
+  port: 3231 + parseInt(process.env.CPU_AFFINITY || '0'),
   reqs: 100
 }
