+++ benchmark/sirun/plugin-bluebird/meta.json
@@ -3,7 +3,7 @@
   "run": "node index.js",
   "run_with_affinity": "bash -c \"taskset -c $CPU_AFFINITY node index.js\"",
   "cachegrind": false,
-  "iterations": 250,
+  "iterations": 100,
   "instructions": true,
   "variants": {
     "control": {
