+++ benchmark/sirun/plugin-http/common.js
@@ -1,6 +1,6 @@
 'use strict'
 
 module.exports = {
-  port: 3031,
+  port: 3031 + parseInt(process.env.CPU_AFFINITY || '0'),
   reqs: 350
 }
