+++ benchmark/sirun/plugin-net/common.js
@@ -1,6 +1,6 @@
 'use strict'
 
 module.exports = {
-  port: 3033,
+  port: 3131 + parseInt(process.env.CPU_AFFINITY || '0'),
   reqs: 100
 }
