+++ benchmark/sirun/run-all-variants.js
@@ -2,79 +2,18 @@
 
 'use strict'
 
-const childProcess = require('child_process')
 const fs = require('fs')
 const path = require('path')
-const readline = require('readline')
+const { exec, getStdio } = require('./run-util')
 
 process.env.DD_TRACE_TELEMETRY_ENABLED = 'false'
 
-function exec (...args) {
-  return new Promise((resolve, reject) => {
-    const proc = childProcess.spawn(...args)
-    streamAddVersion(proc.stdout)
-    proc.on('error', reject)
-    proc.on('exit', (code) => {
-      if (code === 0) {
-        resolve()
-      } else {
-        reject(new Error('Process exited with non-zero code.'))
-      }
-    })
-  })
-}
+require('./squash-affinity')
 
 const metaJson = require(path.join(process.cwd(), 'meta.json'))
-
-if (process.env.ENABLE_AFFINITY) {
-  squashAffinity(metaJson)
-
-  if (metaJson.variants) {
-    const variants = metaJson.variants
-
-    for (const variantName in variants) {
-      const variant = variants[variantName]
-      squashAffinity(variant)
-    }
-  }
-}
-
-function squashAffinity (obj) {
-  if (obj.run_with_affinity) {
-    obj.run = obj.run_with_affinity
-    delete obj.run_with_affinity
-  }
-
-  if (obj.setup_with_affinity) {
-    obj.setup = obj.setup_with_affinity
-    delete obj.setup_with_affinity
-  }
-}
-
-fs.writeFileSync(path.join(process.cwd(), 'meta-temp.json'), JSON.stringify(metaJson, null, 2))
-
 const env = Object.assign({}, process.env, { DD_TRACE_STARTUP_LOGS: 'false' })
 
-function streamAddVersion (input) {
-  input.rl = readline.createInterface({ input })
-  input.rl.on('line', function (line) {
-    try {
-      const json = JSON.parse(line.toString())
-      json.nodeVersion = process.versions.node
-      // eslint-disable-next-line no-console
-      console.log(JSON.stringify(json))
-    } catch (e) {
-      // eslint-disable-next-line no-console
-      console.log(line)
-    }
-  })
-}
-
-function getStdio () {
-  return ['inherit', 'pipe', 'inherit']
-}
-
-(async () => {
+;(async () => {
   try {
     if (metaJson.variants) {
       const variants = metaJson.variants
