+++ benchmark/sirun/startup/meta.json
@@ -3,7 +3,7 @@
   "run": "node startup-test.js",
   "run_with_affinity": "bash -c \"taskset -c $CPU_AFFINITY node startup-test.js\"",
   "cachegrind": false,
-  "iterations": 40,
+  "iterations": 20,
   "instructions": true,
   "variants": {
     "control": {
